code/03/vscode-chatflow/requirements.txt
azure-search-documents==11.4.0b8
azure-storage-blob 
azure-identity  